requirements.txt
fastapi~=0.103.1
pydantic~=2.3.0
starlette~=0.27.0
slowapi~=0.1.8
sqlalchemy~=2.0.21
pymysql~=1.1.0
uvicorn~=0.23.2
loguru~=0.7.2
bcrypt~=4.0.1
python-jose~=3.3.0